+++ test/pay/stripe/customer_test.rb
@@ -490,6 +490,14 @@ class Pay::Stripe::CustomerTest < ActiveSupport::TestCase
     @pay_customer.create_meter_event(:api_request, payload: {value: 1})
   end
 
+  test "stripe retry_past_due_subscriptions! pays open invoices of past_due subscriptions" do
+    @pay_customer.update!(processor_id: "cus_1234")
+    pay_subscriptions(:stripe).update!(status: :past_due)
+    ::Stripe::Invoice.expects(:list).with({subscription: "sub_1", status: :open, expand: ["data.payments"]}, {}).returns(::Stripe::ListObject.construct_from(object: "list", has_more: false, data: []))
+
+    @pay_customer.retry_past_due_subscriptions!
+  end
+
   private
 
   def payment_method